// src/yeying/api/common/code.ts
export enum ServiceCodeEnum {
  /** SERVICE_CODE_UNKNOWN - 未知编码，是一种占位符 */
  SERVICE_CODE_UNKNOWN = 0,
  /** SERVICE_CODE_NODE - 网络节点供应商 */
  SERVICE_CODE_NODE = 1,
  /** SERVICE_CODE_WAREHOUSE - 仓储服务供应商 */
  SERVICE_CODE_WAREHOUSE = 2,
  /** SERVICE_CODE_AGENT - 智能体供应商 */
  SERVICE_CODE_AGENT = 3,
  /** SERVICE_CODE_MCP - 模型上下文供应商 */
  SERVICE_CODE_MCP = 4,
  UNRECOGNIZED = -1,
}
// ...
export enum ApiCodeEnum {
  /** API_CODE_UNKNOWN - 未知编码，是一种占位符 */
  API_CODE_UNKNOWN = 0,
  /** API_CODE_USER - 用户服务 */
  API_CODE_USER = 1,
  /** API_CODE_IDENTITY - 身份服务 */
  API_CODE_IDENTITY = 2,
  /** API_CODE_LLM_SERVICE - 大模型服务API */
  API_CODE_LLM_SERVICE = 3,
  /** API_CODE_LLM_PROVIDER - 大模型供应商API */
  API_CODE_LLM_PROVIDER = 4,
  /** API_CODE_ASSET_SERVICE - 资产服务API */
  API_CODE_ASSET_SERVICE = 5,
  /** API_CODE_ASSET_BLOCK - 资产块API */
  API_CODE_ASSET_BLOCK = 6,
  /** API_CODE_ASSET_LINK - 资产链接API */
  API_CODE_ASSET_LINK = 7,
  /** API_CODE_ASSET_NAMESPACE - 资产命名空间API */
  API_CODE_ASSET_NAMESPACE = 8,
  /** API_CODE_ASSET_RECYCLE - 资产回收站API */
  API_CODE_ASSET_RECYCLE = 9,
  /** API_CODE_CERTIFICATE - 证书服务 */
  API_CODE_CERTIFICATE = 10,
  /** API_CODE_STORAGE - 存储服务 */
  API_CODE_STORAGE = 11,
  /** API_CODE_APPLICATION - 应用仓库 */
  API_CODE_APPLICATION = 12,
  /** API_CODE_EVENT - 事件服务 */
  API_CODE_EVENT = 13,
  /** API_CODE_INVITATION - 邀请码服务 */
  API_CODE_INVITATION = 14,
  /** API_CODE_SERVICE - 服务接口 */
  API_CODE_SERVICE = 15,
  /** API_CODE_RAG - RAG 接口 */
  API_CODE_RAG = 16,
  UNRECOGNIZED = -1,
}