// src/yeying/api/service/service.ts
export interface CreateServiceResponse {
  header: MessageHeader | undefined;
  body: CreateServiceResponseBody | undefined;
}

export interface CreateServiceResponseBody {
  status: ResponseStatus | undefined;
  service: ServiceMetadata | undefined;
}

/** detail */
export interface DetailServiceRequest {
  header: MessageHeader | undefined;
  body: DetailServiceRequestBody | undefined;
}

export interface DetailServiceRequestBody {
  /** 服务身份 did */
  did: string;
  /** 服务版本 */
  version: number;
}

export interface DetailServiceResponse {
  header: MessageHeader | undefined;
  body: DetailServiceResponseBody | undefined;
}

export interface DetailServiceResponseBody {
  status: ResponseStatus | undefined;
  service: ServiceMetadata | undefined;
}

/** search */
export interface SearchServiceRequest {
  header: MessageHeader | undefined;
  body: SearchServiceRequestBody | undefined;
}

export interface SearchServiceRequestBody {
  condition: SearchServiceCondition | undefined;
  page: RequestPage | undefined;
}

export interface SearchServiceCondition {
  /** 服务类型【参数可选】 */
  code: ServiceCodeEnum;
  /** 服务所有者【参数可选】 */
  owner: string;
  /** 服务名称【参数可选】 */
  name: string;
  /** 输入[服务所有者、服务名称]关键字搜索【参数可选】 */
  keyword: string;
}

export interface SearchServiceResponse {
  header: MessageHeader | undefined;
  body: SearchServiceResponseBody | undefined;
}

export interface SearchServiceResponseBody {
  status: ResponseStatus | undefined;
  services: ServiceMetadata[];
  page: ResponsePage | undefined;
}

/** delete */
export interface DeleteServiceRequest {
  header: MessageHeader | undefined;
  body: DeleteServiceRequestBody | undefined;
}

export interface DeleteServiceRequestBody {
  did: string;
  version: number;
}

export interface DeleteServiceResponse {
  header: MessageHeader | undefined;
  body: DeleteServiceResponseBody | undefined;
}

export interface DeleteServiceResponseBody {
  status: ResponseStatus | undefined;
}
// ...
export const CreateServiceResponse: MessageFns<CreateServiceResponse> = {
  encode(message: CreateServiceResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      CreateServiceResponseBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateServiceResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateServiceResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = CreateServiceResponseBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateServiceResponse {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? CreateServiceResponseBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: CreateServiceResponse): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = CreateServiceResponseBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateServiceResponse>, I>>(base?: I): CreateServiceResponse {
    return CreateServiceResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateServiceResponse>, I>>(object: I): CreateServiceResponse {
    const message = createBaseCreateServiceResponse();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? CreateServiceResponseBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};

function createBaseCreateServiceResponseBody(): CreateServiceResponseBody {
  return { status: undefined, service: undefined };
}

export const CreateServiceResponseBody: MessageFns<CreateServiceResponseBody> = {
  encode(message: CreateServiceResponseBody, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== undefined) {
      ResponseStatus.encode(message.status, writer.uint32(10).fork()).join();
    }
    if (message.service !== undefined) {
      ServiceMetadata.encode(message.service, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateServiceResponseBody {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateServiceResponseBody();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.status = ResponseStatus.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.service = ServiceMetadata.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateServiceResponseBody {
    return {
      status: isSet(object.status) ? ResponseStatus.fromJSON(object.status) : undefined,
      service: isSet(object.service) ? ServiceMetadata.fromJSON(object.service) : undefined,
    };
  },

  toJSON(message: CreateServiceResponseBody): unknown {
    const obj: any = {};
    if (message.status !== undefined) {
      obj.status = ResponseStatus.toJSON(message.status);
    }
    if (message.service !== undefined) {
      obj.service = ServiceMetadata.toJSON(message.service);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateServiceResponseBody>, I>>(base?: I): CreateServiceResponseBody {
    return CreateServiceResponseBody.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateServiceResponseBody>, I>>(object: I): CreateServiceResponseBody {
    const message = createBaseCreateServiceResponseBody();
    message.status = (object.status !== undefined && object.status !== null)
      ? ResponseStatus.fromPartial(object.status)
      : undefined;
    message.service = (object.service !== undefined && object.service !== null)
      ? ServiceMetadata.fromPartial(object.service)
      : undefined;
    return message;
  },
};
// ...
export const DeleteServiceResponse: MessageFns<DeleteServiceResponse> = {
  encode(message: DeleteServiceResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.header !== undefined) {
      MessageHeader.encode(message.header, writer.uint32(10).fork()).join();
    }
    if (message.body !== undefined) {
      DeleteServiceResponseBody.encode(message.body, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DeleteServiceResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDeleteServiceResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header = MessageHeader.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.body = DeleteServiceResponseBody.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DeleteServiceResponse {
    return {
      header: isSet(object.header) ? MessageHeader.fromJSON(object.header) : undefined,
      body: isSet(object.body) ? DeleteServiceResponseBody.fromJSON(object.body) : undefined,
    };
  },

  toJSON(message: DeleteServiceResponse): unknown {
    const obj: any = {};
    if (message.header !== undefined) {
      obj.header = MessageHeader.toJSON(message.header);
    }
    if (message.body !== undefined) {
      obj.body = DeleteServiceResponseBody.toJSON(message.body);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DeleteServiceResponse>, I>>(base?: I): DeleteServiceResponse {
    return DeleteServiceResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DeleteServiceResponse>, I>>(object: I): DeleteServiceResponse {
    const message = createBaseDeleteServiceResponse();
    message.header = (object.header !== undefined && object.header !== null)
      ? MessageHeader.fromPartial(object.header)
      : undefined;
    message.body = (object.body !== undefined && object.body !== null)
      ? DeleteServiceResponseBody.fromPartial(object.body)
      : undefined;
    return message;
  },
};

function createBaseDeleteServiceResponseBody(): DeleteServiceResponseBody {
  return { status: undefined };
}

export const DeleteServiceResponseBody: MessageFns<DeleteServiceResponseBody> = {
  encode(message: DeleteServiceResponseBody, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== undefined) {
      ResponseStatus.encode(message.status, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DeleteServiceResponseBody {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDeleteServiceResponseBody();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.status = ResponseStatus.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DeleteServiceResponseBody {
    return { status: isSet(object.status) ? ResponseStatus.fromJSON(object.status) : undefined };
  },

  toJSON(message: DeleteServiceResponseBody): unknown {
    const obj: any = {};
    if (message.status !== undefined) {
      obj.status = ResponseStatus.toJSON(message.status);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DeleteServiceResponseBody>, I>>(base?: I): DeleteServiceResponseBody {
    return DeleteServiceResponseBody.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DeleteServiceResponseBody>, I>>(object: I): DeleteServiceResponseBody {
    const message = createBaseDeleteServiceResponseBody();
    message.status = (object.status !== undefined && object.status !== null)
      ? ResponseStatus.fromPartial(object.status)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}