// src/yeying/api/common/model.ts
export interface ServiceMetadata {
  /** 服务所有者 */
  owner: string;
  /** 服务所在网络 */
  network: string;
  /** 服务所在网络地址 */
  address: string;
  /** 服务身份 */
  did: string;
  /** 服务版本 */
  version: number;
  /** 服务名称 */
  name: string;
  /** 服务描述 */
  description: string;
  /** 服务编码 */
  code: ServiceCodeEnum;
  /** 服务apis */
  apiCodes: ApiCodeEnum[];
  /** 服务代理 */
  proxy: string;
  /** 服务grpc地址 */
  grpc: string;
  /** 服务头像 */
  avatar: string;
  /** 服务创建时间点 */
  createdAt: string;
  /** 服务修改时间点 */
  updatedAt: string;
  /** 签名 */
  signature: string;
  /** 代码包下载路径 */
  codePackagePath: string;
}